Fix for Incorrectly Showing Satellite Icon
Issue: When slot mapping with a new subscription ID occurs, the
Telephony framework reports the CarrierRoamingNtnMode value of the
previous subscription to all clients registered for the telephony
callback.

Solution: To resolve this issue, reset the cache of the
CarrierRoamingNtnMode value of the previous subscription when it is
removed.

Change-Id: I37c75cfae49f7e235eeb2181361ec93f245cd564
CRs-Fixed: 4127523