Fix for unbinding a service repeatedly
In certain cases, when there is a binding action driven by a carrier
configuration change that hasn't started, the unbinding service is
triggered by a device configuration change. This causes the service to
attempt to unbind a service that hasn't been started yet, leading to an
exception and resulting in the process exiting unexpectedly.

To address this issue, ensure that the variable mServiceConnection is
set to null once the unbinding process is complete. This prevents any
duplicate unbinding actions from occurring. Additionally, we’ve added a
step to unlink the death recipient to avoid any excessive reference
count issues in Java.

Change-Id: I5b15945f50b7f2d247c625157988feb89e5de7dc
CRs-Fixed: 4258188